# docs/swagger.yaml
definitions:
  github_com_bsonger_devflow-verify-service_pkg_model.Application:
    properties:
      active_manifest_id:
        type: string
      active_manifest_name:
        type: string
      config_maps:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.ConfigMap'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      envs:
        additionalProperties:
          items:
            $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.EnvVar'
          type: array
        type: object
      id:
        type: string
      internet:
        $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Internet'
      name:
        type: string
      project_id:
        type: string
      project_name:
        type: string
      replica:
        type: integer
      repo_url:
        type: string
      service:
        $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Service'
      status:
        type: string
      type:
        $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.ReleaseType'
      updated_at:
        type: string
    type: object
  github_com_bsonger_devflow-verify-service_pkg_model.ConfigMap:
    properties:
      files_path:
        additionalProperties:
          type: string
        type: object
      mount_path:
        type: string
      name:
        type: string
    type: object
  github_com_bsonger_devflow-verify-service_pkg_model.Configuration:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      files:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.File'
        type: array
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_bsonger_devflow-verify-service_pkg_model.EnvVar:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  github_com_bsonger_devflow-verify-service_pkg_model.File:
    properties:
      content:
        type: string
      name:
        type: string
    type: object
  github_com_bsonger_devflow-verify-service_pkg_model.Intent:
    properties:
      application_id:
        type: string
      application_name:
        type: string
      attempt_count:
        type: integer
      branch:
        type: string
      claimed_at:
        type: string
      claimed_by:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      env:
        type: string
      external_ref:
        type: string
      id:
        type: string
      job_id:
        type: string
      job_type:
        type: string
      kind:
        $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.IntentKind'
      last_error:
        type: string
      lease_expires_at:
        type: string
      manifest_id:
        type: string
      manifest_name:
        type: string
      message:
        type: string
      repo_url:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
      status:
        $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.IntentStatus'
      updated_at:
        type: string
    type: object
  github_com_bsonger_devflow-verify-service_pkg_model.IntentKind:
    enum:
    - build
    - release
    type: string
    x-enum-varnames:
    - IntentKindBuild
    - IntentKindRelease
  github_com_bsonger_devflow-verify-service_pkg_model.IntentStatus:
    enum:
    - Pending
    - Running
    - Succeeded
    - Failed
    type: string
    x-enum-varnames:
    - IntentPending
    - IntentRunning
    - IntentSucceeded
    - IntentFailed
  github_com_bsonger_devflow-verify-service_pkg_model.Internet:
    enum:
    - internal
    - external
    type: string
    x-enum-varnames:
    - Internal
    - External
...
  github_com_bsonger_devflow-verify-service_pkg_model.Port:
    properties:
      name:
        type: string
      port:
        type: integer
      target_port:
        type: integer
    type: object
...
  github_com_bsonger_devflow-verify-service_pkg_model.ReleaseType:
    enum:
    - normal
    - canary
    - blue-green
    type: string
    x-enum-varnames:
    - Normal
    - Canary
    - BlueGreen
  github_com_bsonger_devflow-verify-service_pkg_model.Service:
    properties:
      ports:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Port'
        type: array
    type: object
...
  pkg_api.CreateResponse:
    properties:
      execution_intent_id:
        type: string
      id:
        type: string
    type: object
  pkg_api.UpdateActiveManifestRequest:
    properties:
      manifest_id:
        type: string
    required:
    - manifest_id
    type: object
...
info:
  contact: {}
paths:
  /api/v1/applications:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Application'
            type: array
      summary: 获取应用列表
      tags:
      - Application
    post:
      consumes:
      - application/json
      description: 创建一个新的应用
      parameters:
      - description: Application Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Application'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api.CreateResponse'
      summary: 创建应用
      tags:
      - Application
  /api/v1/applications/{id}:
    delete:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除应用
      tags:
      - Application
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Application'
      summary: 获取应用
      tags:
      - Application
    put:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Application Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Application'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新应用
      tags:
      - Application
  /api/v1/applications/{id}/active_manifest:
    patch:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Active Manifest Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_api.UpdateActiveManifestRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新应用的 Active Manifest
      tags:
      - Application
  /api/v1/configurations:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Configuration'
            type: array
      summary: 获取配置列表
      tags:
      - Configuration
    post:
      consumes:
      - application/json
      description: 创建一个新的配置
      parameters:
      - description: Configuration Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Configuration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api.CreateResponse'
      summary: 创建配置
      tags:
      - Configuration
  /api/v1/configurations/{id}:
    delete:
      parameters:
      - description: Configuration ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除配置
      tags:
      - Configuration
    get:
      parameters:
      - description: Configuration ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Configuration'
      summary: 获取配置
      tags:
      - Configuration
    put:
      parameters:
      - description: Configuration ID
        in: path
        name: id
        required: true
        type: string
      - description: Configuration Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Configuration'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新配置
      tags:
      - Configuration
  /api/v1/intents:
    get:
      description: 按 kind、status、resource、application 等维度查询 execution intents
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Intent'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取执行意图列表
      tags:
      - Intent
  /api/v1/intents/{id}:
    get:
      parameters:
      - description: Intent ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-verify-service_pkg_model.Intent'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取执行意图
      tags:
      - Intent
...
swagger: "2.0"